requirements.txt
Django==5.2.6
psycopg2-binary==2.9.10
django-allauth==65.11.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dj-rest-auth==7.0.1
django-autoslug==1.9.9
django-phonenumber-field==8.1.0
phonenumbers==9.0.3
django-countries==7.6.1
phonenumbers==9.0.3
pillow==11.3.0
channels==4.3.1
daphne==4.2.1